Major restructure of the code. Components are separated into modules.
The core simulation (which has no dependencies) now lives in core/
The Qt GUI (depends on core) lives in gui/
The GLUT interface (GLView.h|cpp) is in the root dir with the new main.cpp.

If Qt is found the GLUT moduleis ignored, else the Qt module is ignored and GLUT is used.